Turn libxslt into a submodule.
The original import apparently was from libxslt v1.1.28, so that's what I'm
adding as a submodule as well.  There was a slight discrepancy in the
xsltwin32config.h causing the HAVE_MODULES flag to be always set.  I opted
to set that flahg in the libxslt.gyp file instead, so that we don't have to
patch the upstream sources.