Implement arbitrary extension functions written in JavaScript.
For asynchroneous calls, the callbacks are performed on the main event loop
while the worker thread is blocked on a mutex.  At the moment, only boolean,
numeric and string values can be passed between the XSLT processor and the
JavaScript function.  Other types would be highly desirable, but might
require a lot more work.