Make builds on AppVeyor succeed
To achieve that, we have to use a newer version of libxmljs-mt which avoids
defining the snprintf function as a macro and does ddl-export its api.

To make the nodejs_version setting have any effect we also need to install
node ourself.  See http://www.appveyor.com/docs/lang/nodejs-iojs for that.

We need to define LIBXSLT_DEFAULT_PLUGINS_PATH to avoid an unresolved
symbol, but we may define it to be NULL which seems to be what
`deps/libxslt/wind32/configure.js` would be doing as well.